[FLINK-15179] Make Kubernetes Session CLI use the ExecutorCLI
